Add MinGW32 support

+++ src/engine/audio/include/halley/audio/vorbis_dec.h
@@ -30,7 +30,7 @@ struct OggVorbis_File;
 #if defined(_MSC_VER) || defined(__clang__)
 using OggOffsetType = int64_t;
 #else
-using OggOffsetType = long int;
+using OggOffsetType = long long;
 #endif
 
 namespace Halley {
